Adjustments to tests and references
Per discussion with Sezzle's awesome developer support, added
appropriate comments in integration tests. Unfortunately we can't
completely automate a checkout process at this time...so it'll involve
firing off a test, completing steps in a browser and then finishing up
another nunit test. This is working as expected at this time aside from
the order items not showing in the Sezzle Merchant Dashboard. Currently
in discussion w / support to see what's happening here and if this is
expected behaviour or not.