Misc. updates
Removed rate limit related items as Sezzle doesn't currently rate limit.
Also, adjusted ExecuteAsync extension method to support a scenario where
an empty body is returned. In those scenarios, we'll set the
HttpStatusCode on the Response as well as the StatusCode and Message
values to match the values returned in the headers.